Reduce complexity of verify_can_exec funcs
This applies to most action types. The original way that the verify_can_exec
functions were written predates a nice 'filesys' object to wrap things like
os.access(). Now that the filesys can handle high level operations, there's no
need for every verification function to break down checks into multiple locally
defined functions. Also adds a bit of extra tooling to the abstract filesys
type to support this behavior for the common case of checking writability of a
file or its only existing ancestor.
