+++ salve/action/copy/directory.py
@@ -5,23 +5,6 @@ from salve.context import ExecutionContext
 
 
 class DirCopyAction(CopyAction):
-    """
-    An action to copy a directory tree.
-    """
-    def __init__(self, src, dst, file_context):
-        """
-        DirCopyAction constructor.
-
-        Args:
-            @src
-            Source path.
-            @dst
-            Destination path.
-            @file_context
-            The FileContext.
-        """
-        CopyAction.__init__(self, src, dst, file_context)
-
     def __str__(self):
         return ("DirCopyAction(src=" + str(self.src) + ",dst=" +
                 str(self.dst) + ",context=" + repr(self.file_context) + ")")
@@ -31,30 +14,13 @@ class DirCopyAction(CopyAction):
         Check to ensure that execution can proceed without errors.
         Ensures that the the target directory is writable.
         """
-        # transition to the action verification phase,
-        # confirming execution will work
         ExecutionContext().transition(ExecutionContext.phases.VERIFICATION)
-
-        def readable_source():
-            """
-            Checks if the source is a readable and traversable directory. If
-            not, then it will be impossible to view and copy its contents.
-            """
-            return filesys.access(self.src,
-                                  access_codes.R_OK | access_codes.X_OK)
-
-        def writable_target():
-            """
-            Checks if the target is in a writable directory.
-            """
-            return filesys.access(paths.dirname(self.dst),
-                                  access_codes.W_OK)
 
         logstr = ('DirCopy: Checking source is readable + traversable, ' +
                   '{0}'.format(self.dst))
         logger.info('{0}: {1}'.format(self.file_context, logstr))
 
-        if not readable_source():
+        if not filesys.access(self.src, access_codes.R_OK | access_codes.X_OK):
             return self.verification_codes.UNREADABLE_SOURCE
 
         logger.info(
@@ -62,7 +28,7 @@ class DirCopyAction(CopyAction):
                 self.file_context, self.dst)
             )
 
-        if not writable_target():
+        if not filesys.access(paths.dirname(self.dst), access_codes.W_OK):
             return self.verification_codes.UNWRITABLE_TARGET
 
         return self.verification_codes.OK
@@ -83,7 +49,6 @@ class DirCopyAction(CopyAction):
             logger.warn('{0}: {1}'.format(self.file_context, logstr))
             return
 
-        # transition to the execution phase
         ExecutionContext().transition(ExecutionContext.phases.EXECUTION)
 
         logstr = ('Performing Directory Copy \"%s\" -> \"%s\"' %
