+++ salve/action/create/directory.py
@@ -1,8 +1,6 @@
 from salve import logger
-from salve.action.base import Action
 from salve.action.create.base import CreateAction
 
-from salve.filesys import access_codes
 from salve.context import ExecutionContext
 
 
@@ -10,19 +8,6 @@ class DirCreateAction(CreateAction):
     """
     An action to create a directory.
     """
-    def __init__(self, dst, file_context):
-        """
-        DirCreateAction constructor.
-
-        Args:
-            @dst
-            Destination path.
-            @file_context
-            The FileContext.
-        """
-        Action.__init__(self, file_context)
-        self.dst = dst
-
     def __str__(self):
         return ("DirCreateAction(dst=" + self.dst + ",context=" +
                 repr(self.file_context) + ")")
@@ -31,16 +16,7 @@ class DirCreateAction(CreateAction):
         """
         Checks if the target dir already exists, or if its parent is writable.
         """
-        # transition to the action verification phase,
-        # confirming execution will work
         ExecutionContext().transition(ExecutionContext.phases.VERIFICATION)
-
-        def writable_target():
-            """
-            Checks if the target is in a writable directory.
-            """
-            ancestor = filesys.get_existing_ancestor(self.dst)
-            return filesys.access(ancestor, access_codes.W_OK)
 
         logstr = 'DirCreate: Checking if target exists, \"%s\"' % self.dst
         logger.info('{0}: {1}'.format(self.file_context, logstr))
@@ -52,7 +28,7 @@ class DirCreateAction(CreateAction):
         logstr = 'DirCreate: Checking target is writable, \"%s\"' % self.dst
         logger.info('{0}: {1}'.format(self.file_context, logstr))
 
-        if not writable_target():
+        if not filesys.writable_path_or_ancestor(self.dst):
             return self.verification_codes.UNWRITABLE_TARGET
 
         return self.verification_codes.OK
@@ -69,7 +45,6 @@ class DirCreateAction(CreateAction):
             logger.warn('{0}: {1}'.format(self.file_context, logstr))
             return
 
-        # transition to the execution phase
         ExecutionContext().transition(ExecutionContext.phases.EXECUTION)
 
         logstr = 'Performing Directory Creation of \"%s\"' % self.dst
