+++ salve/action/create/file.py
@@ -1,7 +1,6 @@
-from salve import logger, paths
+from salve import logger
 from salve.action.create.base import CreateAction
 
-from salve.filesys import access_codes
 from salve.context import ExecutionContext
 
 
@@ -9,19 +8,6 @@ class FileCreateAction(CreateAction):
     """
     An action to create a single file.
     """
-    def __init__(self, dst, file_context):
-        """
-        FileCreateAction constructor.
-
-        Args:
-            @dst
-            Destination path.
-            @file_context
-            The FileContext.
-        """
-        CreateAction.__init__(self, file_context)
-        self.dst = dst
-
     def __str__(self):
         return ("FileCreateAction(dst=" + self.dst +
                 ",context=" + repr(self.file_context) + ")")
@@ -31,33 +17,12 @@ class FileCreateAction(CreateAction):
         Ensures that the target file exists and is writable, or that
         it does not exist and is in a writable directory.
         """
-        # transition to the action verification phase,
-        # confirming execution will work
         ExecutionContext().transition(ExecutionContext.phases.VERIFICATION)
-
-        def writable_target():
-            """
-            Checks if the target is in a writable directory.
-            """
-            if filesys.access(self.dst, access_codes.W_OK):
-                return True
-            if filesys.access(self.dst, access_codes.F_OK):
-                return False
-            # file is now known not to exist
-            assert not filesys.exists(self.dst)
-
-            parent = paths.dirname(self.dst)
-            if filesys.access(parent, access_codes.W_OK):
-                return True
-
-            # the file is doesn't exist and the containing dir is
-            # not writable or doesn't exist
-            return False
 
         logstr = 'FileCreate: Checking target is writable, \"%s\"' % self.dst
         logger.info('{0}: {1}'.format(self.file_context, logstr))
 
-        if not writable_target():
+        if not filesys.writable_path_or_ancestor(self.dst):
             return self.verification_codes.UNWRITABLE_TARGET
 
         return self.verification_codes.OK
@@ -75,7 +40,6 @@ class FileCreateAction(CreateAction):
             logger.warn('{0}: {1}'.format(self.file_context, logstr))
             return
 
-        # transition to the execution phase
         ExecutionContext().transition(ExecutionContext.phases.EXECUTION)
 
         logstr = 'Performing File Creation of \"%s\"' % self.dst
